refactor: 增加搜索结果排序 & 优化搜索

+++ src/main/java/com/rubiks/backendoasis/model/PapersWithSize.java
@@ -8,8 +8,8 @@ import java.util.List;
 @Data
 public class PapersWithSize {
     private List<PaperEntity> papers;
-    private int size;
-    public PapersWithSize(List<PaperEntity> papers, int size) {
+    private long size;
+    public PapersWithSize(List<PaperEntity> papers, long size) {
         this.papers = papers;
         this.size = size;
     }
